add a minimal test frontend index.html (#42)
Adds a minimal index.html front end that can pull data through the API service using JavaScript. By default it uses the API key "abc123" and loads the table at route /api/dataset, but both the API and the route can be modified by the user and resubmitted. Additionally, it pulls from the route /api/tables and prints a list of the available tables. It also prints the JSON response as text.