ci(actions): only trigger actions when relevant files change (#95)
This PR makes changes to all of the service-related GH actions (linting,
testing, etc.) to ensure that they only get run when relevant files in
the directory are changed (or the action itself).

Given the re-working of this repo, this will save us waiting
unnecessarily long for irrelevant actions to run when we edit `frontend`
(or any service for that matter).

I also added a `workflow_dispatch` trigger for good measure. 

Closes #84