feat(tests): implement array validation in integration tests (#104)
📦 New CLI dependency [`jq`](https://jqlang.org/) was added to facilitate
this. This can be installed (on MacOS) using `brew install jq`

This PR enhances the integration test script by adding support for
validating object responses, and JSON array responses. It facilitates
checking if specific key-value pairs are found _anywhere_ in the
response.

The previous tests were brittle, and depended on the JSON responses
being in a specific order.

Closes #74

---------

Co-authored-by: CJ Yetman <cj@cjyetman.com>