BREAKING CHANGE: strip out the `api` and `database` services (#117)
Given the latest approved version of the PRD for this project, it was
determined that the database and API services that we had originally
envisioned won't be necessary until a later version of the application.

To keep things stream-lined, @AlexAxthelm @cjyetman and myself have
decided to focus solely on the React front-end in this repository.

Shortly, we will enable continuous deployment of this React application
directly to production, based on the GitFlow branching strategy. This
also benefits from us reducing technical debt and overhead that is not
immediately useful or required for the deployment or serving of the
application.

With that said, given that we do anticipate a future need for a database
and API, we have kept the scaffolding of this available in the
[web-api-poc](https://github.com/RMI/web-api-poc) repository.

Specifially, an identical version of the `api` and `db` services, all
relevant actions, tests, integration tests, `docker` and `.env` set-up
and other related goodies have been preserved in that repository.

Relates to https://github.com/RMI/web-api-poc/pull/56

---------

Co-authored-by: CJ Yetman <cj@cjyetman.com>