Skip deploy if key is missing (#154)
Allow Azure Static Web apps to pass check even if the deploy key is a
missing secret (PR from a forked repo, such as `dependabot`).

Closes #153 

Note, this was tested with a fresh PR in #155. There the actions passed
(green checks) but there was no deployment made and therefore no
deployment comment, but there's no immediate positive indicator that
while the site built successfully, there is no corresponding preview
environment.

To reduce the chances of that happening, I've flagged that the site
should only not deploy if the token is missing if the PR is from
dependabot. If we end up having other bots open PRs, we can add them to
the list, but I think this works for now.

Once we merge this, we should be able to rebase #146 and #147