rename `example.env` to `env.example` (#174)
rename the env file, so that the terminal extension is `.example`,
rather than `.env`. Since this is the only `.env` file in the repo, it
currently is being sourced as the env file automatically when building
the docker image, both with `docker build` and `docker-compose build`.

This change should have no practical effects for `docker compose` users,
since the docker-compose file uses bash variable interpolation to set
the same defaul as defined in the file.

Also aligns filenames with existing README.