Add common docker action for docker build (#176)
Use common docker workflow from `RMI-PACTA`, which handles `hadolint`
and `docker` build and push to GHCR.

Note that this PR is reusing a workflow from `RMI-PACTA/actions`, since
an `RMI/actions` repo doesn't exist yet. Once we've created that, we can
migrate and update the docker actions there.

* Adds `hadolint` check
* pin NPM package version, to pass `hadolint`
* Change docker build process from a `docker compose` centered process
to a standard `docker build`, using `buildx`, which allows builds for
multiple architectures

Note that the workflow this uses for the docker build process doesn't
not create a `latest` tag, but we will have `main` and `production` tags
created automatically (along with `release/*` and `hotfix/*`)

Closes #158