Add unified node GH action caller (#177)
Set up `node` GH workflows (calling other workflows) for simple
management of node-related checks.

Note: this PR will not be able to pass, due to changing the names of
some required actions. as part of approving this PR, we will need to
update branch protection rules to adjust to new action names

Closes #165