Chore/prettier config (#188)
Make the `prettier` config for this project explicit. This PR is open to
discussion on all points. Details for options here:
https://prettier.io/docs/options

Add `prettier` configuration file (and reformat files as needed).

Each commit is usually 1 option set, and if not the default, includes
the diff of running `npm run format` with that config.