Check GH rulesets (#187)
* adds a (non-standard) `.github/rulesets` directory, with JSON exports
of the existing rulesets on this repository.
* Also adds a GH workflow to check that the rulesets in effect in the
repo are the same as defined in `.github/rulesets`

Predecessor to #181 and #178

NOTE: THE GH action runs using the `github.token` for authentication,
which does NOT grant it information about users who can bypass rulesets:

> Note: To prevent leaking sensitive information, the bypass_actors
property is only returned if the user making the API request has write
access to the ruleset.
>
https://docs.github.com/en/enterprise-server@3.13/rest/repos/rules?apiVersion=2022-11-28

To work around this, this action ignores the `bypass_actors` property in
both the remote and local files (filtered using `jq 'del()`)